Rename the fedavg_pers to fedavg_finetune
