+++ examples/fedavg_finetune/fedavg_finetune_MNIST_lenet5_noniid.yml
@@ -18,7 +18,7 @@ clients:
     # personalied model
     # set to be true, each client will load
     # the latest trained personalized model
-    persist_personalized_model: True
+    persist_personalized_model: False
 
 server:
     type: fedavg_personalized
@@ -30,7 +30,7 @@ server:
     # set -1 to do personalized only after the final round
     do_personalization_interval: -1
     # total, participant, nonparticipant
-    do_personalization_group: nonparticipant
+    do_personalization_group: total
 
 data:
     # The training and testing dataset
@@ -55,7 +55,7 @@ trainer:
     type: basic_personalized
 
     # The maximum number of training rounds
-    rounds: 2
+    rounds: 5
 
     # The maximum number of clients running concurrently
     max_concurrency: 2
